relocate logic from recording button to UXCamContext
