+++ client/app/(tabs)/settings/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { View, StyleSheet } from 'react-native';
 import { Stack } from 'expo-router';
 import SettingsItem from '../../../components/SettingsItem';
